mcp_requirements.txt
fastmcp>=3.0.0
requests>=2.28.0
pymupdf>=1.24.0
pytesseract>=0.3.10
python-docx>=1.0.0
openpyxl>=3.1.0
pillow>=10.0.0
